Created tests for the edit and show pages of games.
Show test checks that game attributes are rendered in the show
HTML page. While the edit test confirms that the edit a specific
selected game form is rendered.

Also added find method to the edit controller action to assist with
the edit games test and to assist with editing games going forward.