Game forms have been improved in appearance and styling.
Have introduced additional elements to each of the game form fields making
them have a bigger impact on the form page and other pages such
as where they feature as a partial in the game edit page.

CSS added to previously introduced to form css file, now provides clearer
details for the form and a vertical appearance rather then narrow
horizontal labels and fields as it was previously.

Create button styled to be similar to header and footer of the app.