Forms improved further with Stimulus and Other Changes
Character counter Stimulus added via Yarn to the Games Form to limit the amount of characters able to be entered in the description field.

Also separately to the games form, moved home search bar code into a partial inside shared folder,
called the partial in the top of the homepage below the navbar and current time.

Tidied up home and application views files.