First step in system tests in the application
Implemented Ferrum and Cuprite gems to assist with initial game system tests
done in Minitest due to difficulties when trying to create RSpec system
tests and also with Selenium causing difficulties to even create tests
that run without outside errors/interference.
Test helper files and test rake file setup to assist with
getting the system tests to the current condition.

System tests are running in Headless Chrome, the browser is loading
however tests may not be completely correct yet, need further
investigation into why they are erroring.

Currently running to in an error locally when running the tests however
will look to resolve this in the next branch commit.