Reverted sign up form to previous state and improved
Simple Form reintroduced for sign up form.

Centered sign up form fields in the page and expanded their size to make
it clearer to new users(particularly in case of any sign up errors).

Log in appears on the right of the form fields, working on resolving this.

Improved layout of password prompts from yml to make it clearer as to
the error.