Admin Show Games working and User Admin Creations
The show page for each specific game record is now displayed.

Also have revised the user dashboard to be more in tune with Devise, as
password and password confirmation were in Devise/Sign Up but not
Adminstrate.

Users can now be created in the Admin New User page and they exist within
the frontend also when signing in to a User account.

Posts have been removed from the Users section of the Admin interface
while the focus is currently on syncing the Devise/frontend users with
admin control/approval/creation of users.