Nintendo View Page Refinements and Separate App Tweaks
Removed cards from the Nintendo page due to their effect on the page
layout in particular with their overlapping of the footer, envisage to
add cards to this page again soon. Also now only shows the links in the game titles,
 rather than in aseparate place below the game record and surrounding information.

Cards css has been refined to be much more like the Featured Games on the
homepage, will be revisited in the future when cards themselves are added
again to this page.

Turbo updated in several places to fix the content missing error, with
games the links should now work rather than providing a "Content missing"
error.

Updated calling path for toggle_controller to be just Stimulus.