Navbar converted into Two Separate Stimulus dropdowns
Added new Stimulus dropdown controller code, adjusted placement of existing
index JS controller Stimulus dropdown code and added to this also.

Left and right side of the Navbar have been converted into two types of
Navbar Dropdowns. For the left side of the navbar, this is now open
on a click, and closes on a subsequent click. The right navbar has been
transformed into an open on hover, and then when hovered over again
closes back into the button.

As the buttons are aligning to the far left and right sides of the page,
this will be looked into for the next commit for this branch.

Added css for the dropdown content however I feel this is being
overrided by previously introduced css elsewhere.