Various Navbar CSS Improvements Made and Interactivity Responsiveness Increased
Applied a box-sizing reset to ensure padding and borders are included in the element's total width and height.

Increased padding in header for better spacing.

Non noticeable border included for separation.

Default list styles removed and adjust spacing for the list items, and also links styled within the list items.

Included a media query for increased responsiveness.

Increase to margin for better spacing.

Made several changes to the css for the dropdown content: Lighter red for better visibility of it, increased border radius,
and padding added to for better dropdown spacing