Games Helper and related RSpec Tests for the new Helper Method Created
Moved part of the methodology from the game partial to a new games
helper method relating to the image portion of a game record.

Then created two RSpec tests that test when a game is created, whether
the game has a chosen image attached or not.

Also separately updated non-frontend appearing code for the preview
description in a game form.