Request Specs Updated, Rails Controller Testing Gem added for Extending Request Specs Possibilities
Added rails-controller-testing gem to assist with the route guidance
in request test specs to ensure they direct in the right direction.

Testing locally the specs will fail (however they currently point in the
right direction),this is due to two factors, first
with the games spec, user authentication is required in the current
route the spec path takes. In the second scenario with the user request
spec various partials are opening from Devise and shared partials, in
particular "devise/sessions/new" is being rendered which is the intended
outcome and can later be restyled in the spec to achieve this (the same
applies to the newly added game spec also).