Trix and Action Text Setup into reviews
Initial aim was to setup Trix via Action Text to see how it works within
the application, following the intiial setup following the Rails Guides
it did not seem to make much impact in the frontend (they are however
detailed as being Work in progress), as it did not displayin the created
 confirmation flash page and the index of Reviews page(s).

Following tweaks to the css for Trix, visibility of it has increased,
however it requires further work due to the black background theme
throughout the application affecting its rich text body and buttons above.

Not yet able to edit existing reviews with the Trix text, it will show
the review as being updated however it does not contain the newly added
rich text from the Trix editor.

Trix now displays as part of the Reviews creation process, files can be
added with the attachment button and thanks to ActiveStorageand to new
reviews.