Edit and Show Review Displaying Trix and Trix CSS Updated
Replaced Trix rich text name of "review" with "content" to separate it
from the preexisting use of review in the review MVC's and as part of the
game MVC's.

Added to the private review_params method ":content" to ensure it now
displays in the show and edit Review view pages.

Cut out CSS added to Trix previously and created a new Trix css file and
called in the application.scss file, CSS implemented in this new CSS
partial is working in the form as lots of elements of Trix now appear
in white rather than black which affected visibility of Trix due to the
black background in the application. It also on view pages of the reviews
appeared with a white background, will fix this in the next commit for
this branch.