Comments setup inside Game records
Comments have been setup with their own MVC architecture, and display
inside game records.

Comments created are able to created without being a user (for now) and
will only display on the Game show pages to minimise long lists of popular
games with many comments across the application.

Emphasis is on the comments code be as separate as possible from the game
code to avoid too many overlaps and confusions, and to only have the two
appear in conjunction wherever necessary.

Will implement CSS for the Comments in the future, for both the created
comments appearance and the form to create them.

Turbo fuctionality is not fully realised as it is not displaying newly
created comments live and rather currently requires a page refresh to
display the newly added comments, next commit in this branch should fix
this.