Updated to make sure Real Time Comments Creation and Deletions Occur
Discovered that despite the previous commit having working code to create
comments in real time and without requiring a page reload, it was no longer
working, revised the code and included several new partials to ensure
the real time updates are taking place.

Also updated the routes to add in a new delete action for comments, which
also deletes the comments individually per click of the "Delete" link, courtesy
of a turbo_method as opposed to a standard Rails delete action. Comments
controller also updated accordingly with show and delete action code to
ensure the deletions of comments occur.