Replies to Comments Setup, Path Towards Nested Comments
Following adjustments to the comment model and database schema via a new
migration replies to comments are now possible.

Despite Replies to comments being setup, as a text box will appear so that
a comment can be replied to, however as of this commit it displays upon
page reload and at the bottom of the game show page.

Eventual goal is to achieve nested comments, however for now have achieved
replies to comments, next commit will hope to have a deeper association
between the initial comment and the reply to it.

Have also added block quotes code to eventually also have replies reference
the original comment they are replying to due to a quotation of the original
comment in the reply.