Featured Reviews Setup and Additional Code Revisions
Added new instance for featured_reviews inside the pages controller in
order to render two random (for now) reviews in the application's
homepage. Instructed the homepage to only generate the review's title
and rich text content fields for now, to not overpopulate the homepage
with review fields.

Also tidied up code in several other files.