Converted Navbar to only use Dropdown, Simplified Dropdown Stimulus controller
Navbar hover functionality was not working as intended with the right nav
in the navbar, therefore replaced the hover action with  the dropdown action
from the left nav.

Stimulus dropdown controller code as existing code was too long for what it
was now doing, did not require excessive code explaining actions in depth.
Reduced the controller to be only 35 lines.

Also reran the rails stimulus:install to test the dropdown action was working
as intended after the Stimulus dropdown controller and navbar view changes, which
it is.