Nintendo Page Filtered by Game Franchise Added, Sorts based on Checkbox Ticked
I am looking to introduce a variety of new functions for the application,
this new feature of sorting through existing records based on the game
franchise they belong to (created migration file and migrated it to join it
on to the schema).

The checkbox once ticked (as of now) reloads the page and the intention is
it will generate only the games from the game franchise that has been ticked,
later I would like to avoid the additional wait time of a page reload via using Turbo
for real time record filtering.

Adjusted the routes to acommodate for the new nintendo_games_path in the
_nintendo_content partial, will likely require revision in the future depending
upon action chosen as to how game franchises are filtered.

Css reference to toggle added in application.scss as intend to include a
toggle perhaps with the checkboxes or replacing them later on.